% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_lidarecology.R
\name{lai.raster}
\alias{lai.raster}
\title{Produce a raster map of LAI. The resolution of the raster depends of grain.size choosed on lad.voxel() funtion.}
\usage{
lai.raster(VOXELS_LAD, min = 1, relative.value = NULL)
}
\arguments{
\item{VOXELS_LAD}{3D grid of LAD values (output of lad.voxels() function)}

\item{min}{mix canopy height}

\item{relative.value}{LAI map can be made in percentage of a relative lai value (indicate for effective LAI)}
}
\description{
Produce a raster map of LAI. The resolution of the raster depends of grain.size choosed on lad.voxel() funtion.
}
\examples{
# Get the example laz file
normlas.file = system.file("extdata", "lidar_example.laz", package="packageLAD")

# Calculate LAD from voxelization
VOXELS_LAD.5 = lad.voxels(normlas.file,

#Map using absolute values
lai_raster = lai.raster(VOXELS_LAD.5)
x11();plot(lai_raster)

#map using relative values (\%)
relative.lai_raster = lai.raster(VOXELS_LAD.5, relative.value = lidar.lai)
x11();plot(relative.lai_raster)

}
