% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_lidarecology.R
\name{k.coefficient}
\alias{k.coefficient}
\title{Calculate k coefficient provided a known real LAI and the
calculated LAI}
\usage{
k.coefficient(lidar.lai, real.lai = 6)
}
\arguments{
\item{lidar.lai}{the output from lai() function}

\item{real.lai}{numeric, known real LAI}
}
\description{
Calculate k coefficient provided a known real LAI and the
calculated LAI
}
